requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2

python-dotenv
dj-database-url
 djangorestframework
